+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   See issue https://github.com/comit-network/xmr-btc-swap/issues/652.
 - An issue where the ASB withdraw one bitcoin UTXO at a time instead of the whole balance. 
   See issue https://github.com/comit-network/xmr-btc-swap/issues/662
+- Add the ability to view the swap-cli bitcoin balance and withdraw
+  See issue https://github.com/comit-network/xmr-btc-swap/issues/694
 
 ## [0.8.1] - 2021-08-16